--- airbyte-integrations/connectors/source-braintree-no-code/.dockerignore
@@ -1,6 +1,0 @@
-*
-!Dockerfile
-!main.py
-!source_braintree_no_code
-!setup.py
-!secrets
