--- airbyte-integrations/connectors/source-braintree-no-code/acceptance-test-docker.sh
@@ -1,3 +1,0 @@
-#!/usr/bin/env sh
-
-source "$(git rev-parse --show-toplevel)/airbyte-integrations/bases/connector-acceptance-test/acceptance-test-docker.sh"
