--- airbyte-integrations/connectors/source-braintree-no-code/integration_tests/__init__.py
@@ -1,3 +1,0 @@
-#
-# Copyright (c) 2023 Airbyte, Inc., all rights reserved.
-#
