--- airbyte-integrations/connectors/source-braintree-no-code/integration_tests/abnormal_state.json
@@ -1,11 +1,0 @@
-{
-  "customer_stream": {
-    "created_at": "2222"
-  },
-  "dispute_stream": {
-    "received_date": "2222"
-  },
-  "transaction_stream": {
-    "created_at": "2222"
-  }
-}
