--- airbyte-integrations/connectors/source-braintree-no-code/integration_tests/sample_config.json
@@ -1,7 +1,0 @@
-{
-  "merchant_id": "xcv6yxwmjfrcj8hf",
-  "public_key": "x8z5kh7z2y6q23zf",
-  "private_key": "aef9596be052804b13a7cb0c49df8165",
-  "environment": "Sandbox",
-  "start_date": "2021-01-01T00:00:00Z"
-}
