--- airbyte-integrations/connectors/source-braintree-no-code/integration_tests/sample_state.json
@@ -1,11 +1,0 @@
-{
-  "customer_stream": {
-    "created_at": "2021-08-01"
-  },
-  "dispute_stream": {
-    "received_date": "2021-08-01"
-  },
-  "transaction_stream": {
-    "created_at": "2021-08-01"
-  }
-}
