--- airbyte-integrations/connectors/source-braintree-no-code/integration_tests/spec.json
@@ -1,45 +1,0 @@
-{
-  "documentationUrl": "https://docs.airbyte.com/integrations/sources/braintree",
-  "connectionSpecification": {
-    "title": "Braintree Spec",
-    "type": "object",
-    "properties": {
-      "merchant_id": {
-        "title": "Merchant ID",
-        "description": "The unique identifier for your entire gateway account. See the <a href=\"https://docs.airbyte.com/integrations/sources/braintree\">docs</a> for more information on how to obtain this ID.",
-        "name": "Merchant ID",
-        "type": "string"
-      },
-      "public_key": {
-        "title": "Public Key",
-        "description": "Braintree Public Key. See the <a href=\"https://docs.airbyte.com/integrations/sources/braintree\">docs</a> for more information on how to obtain this key.",
-        "name": "Public Key",
-        "type": "string"
-      },
-      "private_key": {
-        "title": "Private Key",
-        "description": "Braintree Private Key. See the <a href=\"https://docs.airbyte.com/integrations/sources/braintree\">docs</a> for more information on how to obtain this key.",
-        "name": "Private Key",
-        "airbyte_secret": true,
-        "type": "string"
-      },
-      "start_date": {
-        "title": "Start Date",
-        "description": "UTC date and time in the format 2017-01-25T00:00:00Z. Any data before this date will not be replicated.",
-        "name": "Start Date",
-        "examples": ["2020", "2020-12-30", "2020-11-22 20:20:05"],
-        "type": "string",
-        "format": "date-time"
-      },
-      "environment": {
-        "title": "Environment",
-        "description": "Environment specifies where the data will come from.",
-        "name": "Environment",
-        "examples": ["sandbox", "production", "qa", "development"],
-        "enum": ["Development", "Sandbox", "Qa", "Production"],
-        "type": "string"
-      }
-    },
-    "required": ["merchant_id", "public_key", "private_key", "environment"]
-  }
-}
