--- airbyte-integrations/connectors/source-braintree-no-code/requirements.txt
@@ -1,2 +1,0 @@
--e ../../bases/connector-acceptance-test
--e .
