--- airbyte-integrations/connectors/source-braintree-no-code/setup.py
@@ -1,30 +1,0 @@
-#
-# Copyright (c) 2023 Airbyte, Inc., all rights reserved.
-#
-
-
-from setuptools import find_packages, setup
-
-MAIN_REQUIREMENTS = [
-    "airbyte-cdk~=0.1",
-    "braintree~=4.21.0"
-]
-
-TEST_REQUIREMENTS = [
-    "pytest~=6.2",
-    "pytest-mock~=3.6.1",
-    "connector-acceptance-test",
-]
-
-setup(
-    name="source_braintree_no_code",
-    description="Source implementation for Braintree No Code.",
-    author="Airbyte",
-    author_email="contact@airbyte.io",
-    packages=find_packages(),
-    install_requires=MAIN_REQUIREMENTS,
-    package_data={"": ["*.json", "*.yaml", "schemas/*.json", "schemas/shared/*.json"]},
-    extras_require={
-        "tests": TEST_REQUIREMENTS,
-    },
-)
