--- airbyte-integrations/connectors/source-braintree-no-code/source_braintree_no_code/__init__.py
@@ -1,8 +1,0 @@
-#
-# Copyright (c) 2023 Airbyte, Inc., all rights reserved.
-#
-
-
-from .source import SourceBraintreeNoCode
-
-__all__ = ["SourceBraintreeNoCode"]
