--- airbyte-integrations/connectors/source-braintree-no-code/unit_tests/data/customers_ids.txt
@@ -1,6 +1,0 @@
-<search-results>
-    <page-size type="integer">50</page-size>
-    <ids type="array">
-        <item>896865626</item>
-    </ids>
-</search-results>
