--- airbyte-integrations/connectors/source-braintree-no-code/unit_tests/data/subscriptions.txt
@@ -1,5 +1,0 @@
-<search-results>
-    <page-size type="integer">50</page-size>
-    <ids type="array">
-  </ids>
-</search-results>
