--- airbyte-integrations/connectors/source-braintree-no-code/unit_tests/data/transaction_ids.txt
@@ -1,7 +1,0 @@
-<search-results>
-    <page-size type="integer">50</page-size>
-    <ids type="array">
-        <item>2cdqmzsg</item>
-        <item>hncnkbsy</item>
-    </ids>
-</search-results>
