+++ airbyte-integrations/connectors/source-braintree/requirements.txt
@@ -1,3 +1,2 @@
-# This file is autogenerated -- only edit if you know what you are doing. Use setup.py for declaring dependencies.
 -e ../../bases/connector-acceptance-test
 -e .
